added task to main.py, moved station filtering to /processors
